fix unit test & move worker related functions to `workers.go`
